⚡️ Improve performance.

+++ crates/nylon-plugin/src/constants.rs
@@ -3,24 +3,24 @@
 // Plugin method constants
 pub mod methods {
     // Control methods
-    pub const NEXT: usize = 1;
-    pub const END: usize = 2;
-    pub const GET_PAYLOAD: usize = 3;
+    pub const NEXT: u32 = 1;
+    pub const END: u32 = 2;
+    pub const GET_PAYLOAD: u32 = 3;
 
     // Response methods
-    pub const SET_RESPONSE_HEADER: usize = 100;
-    pub const REMOVE_RESPONSE_HEADER: usize = 101;
-    pub const SET_RESPONSE_STATUS: usize = 102;
-    pub const SET_RESPONSE_FULL_BODY: usize = 103;
-    pub const SET_RESPONSE_STREAM_DATA: usize = 104;
-    pub const SET_RESPONSE_STREAM_END: usize = 105;
-    pub const SET_RESPONSE_STREAM_HEADER: usize = 106;
-    pub const READ_RESPONSE_FULL_BODY: usize = 107;
+    pub const SET_RESPONSE_HEADER: u32 = 100;
+    pub const REMOVE_RESPONSE_HEADER: u32 = 101;
+    pub const SET_RESPONSE_STATUS: u32 = 102;
+    pub const SET_RESPONSE_FULL_BODY: u32 = 103;
+    pub const SET_RESPONSE_STREAM_DATA: u32 = 104;
+    pub const SET_RESPONSE_STREAM_END: u32 = 105;
+    pub const SET_RESPONSE_STREAM_HEADER: u32 = 106;
+    pub const READ_RESPONSE_FULL_BODY: u32 = 107;
 
     // Request methods
-    pub const READ_REQUEST_FULL_BODY: usize = 200;
-    pub const READ_REQUEST_HEADER: usize = 201;
-    pub const READ_REQUEST_HEADERS: usize = 202;
+    pub const READ_REQUEST_FULL_BODY: u32 = 200;
+    pub const READ_REQUEST_HEADER: u32 = 201;
+    pub const READ_REQUEST_HEADERS: u32 = 202;
 }
 
 // FFI symbol names
